removed unneccessary files
Major update, moved to WSGI framework

The GUI is now based on the Flask WSGI framework, and is written in Python
HTML is generated using templates
Tables in the GUI are now updated using Ajax calls
Tables are now editable, which avoids using a separate popup
Incorporated in the GUI is a feature called local resources. This allows you to manage a cisco device outside of OpenNaaS using Netconf.